[IMP] base_changeset + base_changeset_tier_validation: Compatibility to one2many, related + summary
This change allows to set a rule_id in changeset always (even in the one2many news).
The "missing" fields are auto-set in raw_changes (they are not in vals of create()) when creating a one2many record.
Get the correct value in the review summary if it is a one2many new.

TT45804